backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.1
pydantic==2.12.5
pydantic-settings>=2.0,<3
python-dotenv==1.2.1
aiohttp==3.11.11
vkpymusic>=3.0
aiogram==3.18.0
yt-dlp>=2024.1.1
